requirements.txt
openai
arxiv
rich
python-dotenv
